Checks & Dummy Account
Description: I added checks to see if the calendar already exists this still needs to change but this is a start. Also added a embedded link so when a calendar gets created or used it shows the link in console. Later we will use this to send to front end.

Reviewed-by: Axel